Removed unnessessary info in item card and reorganized wardrobeItemAddSheet
